Move MySiteCardAndItem items to Item sealed class
