+++ WordPress/src/main/java/org/wordpress/android/ui/mysite/MySiteAdapterDiffCallback.kt
@@ -2,10 +2,10 @@ package org.wordpress.android.ui.mysite
 
 import androidx.recyclerview.widget.DiffUtil
 import org.apache.commons.lang3.NotImplementedException
-import org.wordpress.android.ui.mysite.MySiteCardAndItem.CategoryHeaderItem
 import org.wordpress.android.ui.mysite.MySiteCardAndItem.DomainRegistrationCard
 import org.wordpress.android.ui.mysite.MySiteCardAndItem.DynamicCard.QuickStartDynamicCard
-import org.wordpress.android.ui.mysite.MySiteCardAndItem.ListItem
+import org.wordpress.android.ui.mysite.MySiteCardAndItem.Item.CategoryHeaderItem
+import org.wordpress.android.ui.mysite.MySiteCardAndItem.Item.ListItem
 import org.wordpress.android.ui.mysite.MySiteCardAndItem.QuickActionsCard
 import org.wordpress.android.ui.mysite.MySiteCardAndItem.QuickStartCard
 import org.wordpress.android.ui.mysite.MySiteCardAndItem.SiteInfoCard
