+++ WordPress/src/main/java/org/wordpress/android/ui/mysite/MySiteCategoryItemViewHolder.kt
@@ -2,7 +2,7 @@ package org.wordpress.android.ui.mysite
 
 import android.view.ViewGroup
 import org.wordpress.android.databinding.MySiteCategoryHeaderItemBinding
-import org.wordpress.android.ui.mysite.MySiteCardAndItem.CategoryHeaderItem
+import org.wordpress.android.ui.mysite.MySiteCardAndItem.Item.CategoryHeaderItem
 import org.wordpress.android.ui.utils.UiHelpers
 import org.wordpress.android.util.viewBinding
