+++ WordPress/src/test/java/org/wordpress/android/ui/mysite/SiteItemFixtures.kt
@@ -1,8 +1,8 @@
 package org.wordpress.android.ui.mysite
 
 import org.wordpress.android.R
-import org.wordpress.android.ui.mysite.MySiteCardAndItem.CategoryHeaderItem
-import org.wordpress.android.ui.mysite.MySiteCardAndItem.ListItem
+import org.wordpress.android.ui.mysite.MySiteCardAndItem.Item.CategoryHeaderItem
+import org.wordpress.android.ui.mysite.MySiteCardAndItem.Item.ListItem
 import org.wordpress.android.ui.utils.ListItemInteraction
 import org.wordpress.android.ui.utils.UiString.UiStringRes
 import org.wordpress.android.ui.utils.UiString.UiStringText
